Add support for alternate config file and no database password
These updates are in conjunction with new tests being added to fwknop
project. These changes support the ability to run those tests. The
controller can now take a single argument which is the path to a config
file. Also, the config file allows a new setting so that no password is
set or used for the controller database.